+++ yomishi-core/src/dict/parser/kanji.rs
@@ -19,7 +19,7 @@ pub struct Kanji {
 }
 
 impl FromBank for Kanji {
-    fn parse(r: VecDeque<Value>, format: i32) -> serde_json::Result<Self> {
+    fn parse(r: VecDeque<Value>, format: i64) -> serde_json::Result<Self> {
         (if format == 1 { convert_v1 } else { convert_v3 })(r)
     }
 }
