+++ yomishi-core/src/dict/parser/term_meta.rs
@@ -27,7 +27,7 @@ pub struct Pitch {
 }
 
 impl FromBank for TermMeta {
-    fn parse(r: VecDeque<Value>, _: i32) -> serde_json::Result<Self> {
+    fn parse(r: VecDeque<Value>, _: i64) -> serde_json::Result<Self> {
         convert(r)
     }
 }
